revert unnecessary change
